requirements.txt
discord.py
db-sqlite3
requests